Move more code over
